Print something on restyler errors
I *thought* that there was in implicit exit code check in cram tests.
And that without anything extra, a non-zero exit code would fail the
test, and appending "[n]" to your assertion would assert a specific
(non-zero) exit code. I don't know where I got this, and it doesn't seem
to be the case.

Printing something that will fail tests and/or can be asserted on works.